tablas_biblioteca_migradas_correctamente_1

+++ database/migrations/2019_08_02_050221_crear_tabla_permiso_rol.php
@@ -15,9 +15,9 @@ class CrearTablaPermisoRol extends Migration
     {
         Schema::create('permiso_rol', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('rol_id');
+            $table->unsignedBigInteger('rol_id');
             $table->foreign('rol_id', 'fk_permisorol_rol')->references('id')->on('rol')->onDelete('restrict')->onUpdate('restrict');
-            $table->unsignedInteger('permiso_id');
+            $table->unsignedBigInteger('permiso_id');
             $table->foreign('permiso_id', 'fk_permisorol_usuario')->references('id')->on('permiso')->onDelete('restrict')->onUpdate('restrict');
             $table->timestamps();
         });
